add an alert only on strong similarity, consider shell script errors

+++ conf/converter/convertPdfToHtml.sh
@@ -6,11 +6,21 @@ ctph_file=$3
 
 wget -O "$base_file_name.pdf" "$pdf_url"
 
+if [ $? -ne 0 ]
+then
+	return $?
+fi
+
+pdf2htmlEX --embed-font 0 --process-outline 0 "$base_file_name.pdf"
+
+if [ $? -ne 0 ]
+then
+	return $?
+fi
+
 openssl sha256 "$base_file_name.pdf" | awk '{print $2}' > "$base_file_name.sha256"
 ssdeep -b "$base_file_name.pdf" | tail -n 1 > "$base_file_name.ctp"
 cat "$base_file_name.ctp" >> "$ctph_file"
-
-pdf2htmlEX --embed-font 0 --process-outline 0 "$base_file_name.pdf"
 
 rm "$base_file_name.pdf"
 rm *.woff
